clean up

+++ core/libs/dimg/filters/imgqsort/exposure_detection.cpp
@@ -122,8 +122,8 @@ float ExposureDetector::percent_overexposed()
 
     int normal_pixel = d->image.total() - over_exposed_pixel - demi_over_exposed_pixel;
 
-    return static_cast<float>(static_cast<float>(over_exposed_pixel * d->weight_over_exposure + demi_over_exposed_pixel * d->weight_demi_over_exposure) / 
-                              static_cast<float>(normal_pixel + over_exposed_pixel * d->weight_over_exposure + demi_over_exposed_pixel * d->weight_demi_over_exposure));
+    return static_cast<float>(over_exposed_pixel * d->weight_over_exposure + demi_over_exposed_pixel * d->weight_demi_over_exposure) / 
+           static_cast<float>(normal_pixel + over_exposed_pixel * d->weight_over_exposure + demi_over_exposed_pixel * d->weight_demi_over_exposure);
 }
 
 float ExposureDetector::percent_underexposed()
@@ -134,8 +134,8 @@ float ExposureDetector::percent_underexposed()
 
     int normal_pixel = d->image.total() - under_exposed_pixel - demi_under_exposed_pixel;
 
-    return static_cast<float>(static_cast<float>(under_exposed_pixel * d->weight_under_exposure + demi_under_exposed_pixel * d->weight_demi_under_exposure) / 
-                              static_cast<float>(normal_pixel + under_exposed_pixel * d->weight_under_exposure + demi_under_exposed_pixel * d->weight_demi_under_exposure));
+    return static_cast<float>(under_exposed_pixel * d->weight_under_exposure + demi_under_exposed_pixel * d->weight_demi_under_exposure) / 
+           static_cast<float>(normal_pixel + under_exposed_pixel * d->weight_under_exposure + demi_under_exposed_pixel * d->weight_demi_under_exposure);
 }
 
 int ExposureDetector::count_by_condition(int minVal, int maxVal)
